Reorganise into workspace
This simplifies the dependencies. There's not need to manage cargo
features, etc.
