+++ font/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "profont"
+description = "The ProFont monospace font for use with the embedded-graphics crate."
+version = "0.5.0"
+authors = ["Wesley Moore <wes@wezm.net>"]
+edition = "2018"
+
+documentation = "https://docs.rs/profont"
+repository = "https://github.com/wezm/profont"
+
+readme = "README.md"
+license = "MIT"
+
+keywords = ["bitmap", "monochrome", "font", "monospace"]
+categories = ["rendering", "embedded", "no-std"]
+
+[dependencies]
+embedded-graphics = "0.7.0"
